(Engineering) Speed up compiling for some translation units
* This change removes `#include <sage3basic.h>` from most of the
  changed *.C files in this commit. That header directly or indirectly
  included approximately 1,000,000 lines of C++ code that took around
  35 seconds to compile.

Issue #28

+++ src/frontend/BinaryFormats/GenericSection.C
@@ -827,7 +827,7 @@ SgAsmGenericSection::dumpContainingSections(FILE *f, const std::string &prefix, 
         if (s->isMapped() && rva>=s->get_mappedPreferredRva() && rva<s->get_mappedPreferredRva()+s->get_mappedSize()) {
             rose_addr_t offset = rva - s->get_mappedPreferredRva();
             fprintf(f, "%-*s   is 0x%08" PRIx64 " (%" PRIu64 ") bytes into section [%d] \"%s\"\n",
-                    DUMP_FIELD_WIDTH, prefix.c_str(), offset, offset, s->get_id(), s->get_name()->get_string(true).c_str());
+                    (int)DUMP_FIELD_WIDTH, prefix.c_str(), offset, offset, s->get_id(), s->get_name()->get_string(true).c_str());
         }
     }
 }
